General App: file upload and filtering in DT
The first framework for the general app, including a file upload possibility for .csv and .xlsx datasets. Filtering can be done directly in the DT. The new filtered dataset appears in the tab 'New Data Overview'. 

A design for defining the class of the columns (numeric, character,...) is included but not functionally implemented yet. (Idea is to make sure numerical values are read in as numbers and not as characters)